Start moving toward a callback model for code-completion type-checking
... to reduce and someday eliminate all the places we do double
typechecking.  The goal is to setup any 'interesting' expressions ahead
of time so that if we type-check them at any point - e.g.  during
checking the context - we don't lose the results and try to check them
again.

For now, this is limited to cases where the original expression is not
replaced by pre-checking or CSGen (i.e. we don't handle
UnresolvedDeclRef through this mechanism yet), and we're only using it
for the "parsed expression", which is either the base of a dot
completion, or the name in a function-call or initializer call
expression.  The intent is to move unresolved member completion to the
same model in the future.

Currently, we use the results from the callbacks for 60% of these
completions, and of the 40% we miss, about 80% are simply
UnresolvedDeclRef or similarly obvious cases.

Progress toward rdar://21466394
Incidentally fixes rdar://20962472