[Serialization] Cross-reference imported ObjC methods by selector.
This guards against the names of Objective-C methods changing via
NS_SWIFT_NAME, or the types changing by "fixing" optionality or
adding generic arguments to a collection. In most cases, references
to such methods should just pick up the new definition rather than
resulting in a failure to load.

I'd like to extend this model to initializers, properties, and
subscripts as well; properties and subscripts should use the getter's
selector as a key.

This is currently guarded by a frontend flag,
-enable-experimental-objc-xrefs, and off by default. It does
potentially have problematic implications, such as serialized SIL
having a type change out from under it. We need to think through this
a little more (and add more tests) before turning it on.