[SwiftPGO] Store branch taken counts in CondBranchInst (NFC)
Update the SILBuilder and other relevant bits of SILGen s.t we pass
along branch taken counts whenever we create conditional branches.

There is a lot of unfinished work here: we need to teach ParseSIL,
DeserializeSIL, and the SILOptimizer about profile data. And add a
stable AST hasher. And add tests for it all.

Let's skip all of that for now. Next, we'll wire in some actual profile
data.