[test] Using lit incremental mode for by default
The incremental mode orders test by the last time modified, instead of
by the path, it also touches the files for failed tests, in order for
them to be prioritized on the next run. Using this mode as the default
improves both the CI and local development experience: CI builds
triggered by the commit, that includes changes to the tests, will run
those tests first; locally, it makes a lot of sense to run previously
failed tests first to make sure they've been fixed.