SILGen tests: always build modules and API notes when importing the mock SDK
Previously, only test/SILGen/objc_nonnull_lie_hack.swift would build the
API notes for the 'gizmo' module.  Depending on the order in which tests
ran, the module cache would either contain the 'gizmo' module with
nullability annotations, or without.  If we were unlucky and the module
cache was created by a test that did not build the API notes,
test/SILGen/objc_nonnull_lie_hack.swift failed.

This change also removes 46 instances of '-enable-source-import'
(63 more left in the testsuite).