[Type checker] Add test for non-bridged returns in throwing methods.
A throwing method can only be exposed to Objective-C if we can map the
return type convention, either because the return type is Void or
because it is something that is bridged to an object type (and can
therefore be nil to indicate error). However, this excludes cases
where a type is mapped directly to Objective-C but *also* has a
bridging representation, e.g., a return of type 'Int' or
'UInt32'.

Add a test case from rdar://problem/28035614, which was fixed by my
refactoring of ASTContext::getBridgedToObjC() in
51529ae8880af6d111c216aa51abe7161adb374b.

The test case