[swiftc] COFF does not support PrivateLinkage
Changing to InternalLinkage works under Cygwin, but I think some conditional modification to this PR is needed for other platforms. (or this PR should be moved to LLVM ?)

I tested under Cygwin/MinGW.

The variable and function symbols in object files, generated with PrivateLinkage by LLVM,
are viewable and not internal. One of the symbol name is `Lobjectdestroy`.

I don't know why the letter 'L' is prefixed to the symbol name.

Anyway, it generates duplicate error when linking with other object having same named symbol.

Because the 'objectdestroy' code is not same to all object, ignoring duplicate cann't be the
solution.