Don't drop requirements from potential archetypes if they came from a child
It was possible to create a same-type requirement between a type and
one of its associated types. If one of those associated types had
inferred conformance constraints, we would drop them from the parent
type.

This is not actually correct because then we have no way to recover
metadata for the parent type.

The underlying problem has been present for a long time, however this
particular failure mode was new to Swift 3, when we stopped passing
metadata for associated types.

Fixes <rdar://problem/27018457>.