SIL: Fix parsing of specialized conformances in SIL parser
The previous patch regressed matters in a way that was only caught by
the parse_stdlib tests in check-swift-all.

Basically in SIL parsing we have to always pass a valid generic
environment down to performTypeLocChecking(); the
PartialGenericTypeToArchetypeResolver hack no longer works for SIL
mode.

Doing this properly also allows some code in the SIL parser to be
cleaned up and consolidated in the process.