[test] Remove "RUN: true" lit commands
The lit test runner runs all `RUN:` commands in test files, and fails
the test if the commands exit with a non-zero status. In this sense,
including `RUN: false` in a lit test is an easy way to cause a test to
fail, perhaps as a means of debugging its output at some intermediate
step. On the other hand, there's no purpose in including `RUN: true`.
That command will always pass.

Remove the unneeded lit commands.