run LLDB gtests on macOS via the Xcode build
This change introduces a helper function to construct the
LLDB Xcode options that are now used in two different locations.
On macOS Xcode builds, it builds and runs the lldb-gtest scheme
in the LLDB workspace.  This step will fail the lldb test phase
if any of the gtests fail to compile or generate a test failure.