[SIL Parser] Use lookupDirect/lookupValue rather than lookupQualified.
The SIL parser will always be looking for declarations in a specific
nominal type or module. This is not a semantic "qualified lookup", so
don't use that entry point. Eliminates a hack from the SIL parser;
also, this is the only place where we were turning off
NL_VisitSupertypes, which I'd like to kill.