[CMake] swiftClangImporter should depend on CLANG_TABLEGEN_TARGETS (#4911)
* [CMake] swiftClangImporter should depend on CLANG_TABLEGEN_TARGETS

Since the swiftClangImporter library uses generated headers from the clang tablegen targets it should have an explicit dependency. This fixes a dependency issue building swift in-tree with LLVM & Clang, and will not impact build-script builds because CLANG_TABLEGEN_TARGETS will be empty.

* [CMake] Add a comment explaining CLANG_TABLEGEN_TARGETS

The comment explains why this block doesn't need to be wrapped for non-standalone builds.