Fix swift tests to work with in-tree LLDB
Several of the swift repl tests assume that lldb does not exist next to the swift driver in the binary directory. This patch updates the tests to remove that assumption. This allows the swift tests to correctly pass if LLDB is built in-tree.

These tests are all checking for invocations of the legacy swift repl. If LLDB is in the directory next to swift they fail because it instead invokes the LLDB repl. This patch makes the tests pass by hard linking the swift driver into a temp directory before running those tests so that swift doesn't find LLDB sitting next to itself.