IRGen: Emit superclass information for remote reflection
In order to perform associated type lookups for an inherited
conformance of a derived class, the reflection library needs
to know superclasses of all classes.

Previously we had no way of recovering this information
(short of some kind of reverse lookup in the metadata cache).
Oops!

To avoid causing problems with old/new compiler vs old/new
libswiftRemoteMirrors.dylib, this introduces the new record
in a backward-compatible way: we emit a fake AnyObject
conformance with a fake 'super' associated type to store
the superclass of a class.

Not yet hooked up in the reflection library itself.