[TypeCheckType] Move Setter newValue logic into Options
By moving the is-setter-newValue logic into a TypeCheckOption, we
allow more local reasoning inside of resolve type and can plumb the
special knowledge directly into applyNonEscapingFromContext. This is a
little cleaner, but mainly has the advantage that it will keep future
refactorings of the AST and escapability more localized.

Added test cases for SILGen as well, to make sure we don't crash
(enforces getter/setter escapability parity).